Refactor to make CarouselView its own component
