+++ packages/react/src/components/layout/FocusLayout.tsx
@@ -1,13 +1,11 @@
-import { Participant, Track } from 'livekit-client';
+import { Participant } from 'livekit-client';
 import * as React from 'react';
 import { useMaybeLayoutContext, useLayoutContext } from '../../context';
 import { mergeProps } from '../../utils';
-import { isTrackBundlePinned, TrackBundleFilter, TrackBundle } from '@livekit/components-core';
+import { TrackBundle } from '@livekit/components-core';
 import { ParticipantTile } from '../../prefabs/ParticipantTile';
 import { ParticipantClickEvent } from '@livekit/components-core';
-import { useTracks, useVisualStableUpdate } from '../../hooks';
-import { TrackLoop } from '../TrackLoop';
-import { useSize } from '../../helper';
+import { CarouselView } from './CarouselView';
 
 export interface FocusLayoutContainerProps extends React.HTMLAttributes<HTMLDivElement> {
   trackBundle?: TrackBundle;
@@ -63,42 +61,5 @@ export function FocusLayout({ trackBundle, ...props }: FocusLayoutProps) {
         />
       )}
     </>
-  );
-}
-
-export interface CarouselViewProps extends React.HTMLAttributes<HTMLMediaElement> {
-  filter?: TrackBundleFilter;
-  filterDependencies?: [];
-}
-
-export function CarouselView({ filter, filterDependencies = [], ...props }: CarouselViewProps) {
-  const asideEl = React.useRef(null);
-  const layoutContext = useMaybeLayoutContext();
-  const trackBundles = useTracks([
-    { source: Track.Source.Camera, withPlaceholder: true },
-    { source: Track.Source.ScreenShare, withPlaceholder: false },
-  ]);
-  const filteredTiles = React.useMemo(() => {
-    const tilesWithoutPinned = trackBundles.filter(
-      (tile) => !layoutContext?.pin.state || !isTrackBundlePinned(tile, layoutContext.pin.state),
-    );
-    return filter ? tilesWithoutPinned.filter(filter) : tilesWithoutPinned;
-  }, [filter, layoutContext?.pin.state, trackBundles, ...filterDependencies]);
-
-  let tileCount = 5;
-  const { width, height } = useSize(asideEl);
-  if (height >= width) {
-    const tileHeight = width * 0.5625; // Based on the width calculate the tile hight for a 16/9 tile.
-    tileCount = Math.max(1, Math.floor(height / Math.max(tileHeight, 1)) - 1);
-  } else {
-    const tileWidth = height * 1.777777777777777; // Based on the height calculate the tile hight for a 16/9 tile.
-    tileCount = Math.floor(width / tileWidth);
-  }
-  const sortedTiles = useVisualStableUpdate(filteredTiles, tileCount);
-
-  return (
-    <aside ref={asideEl} {...props}>
-      {props.children ?? <TrackLoop trackBundles={sortedTiles} />}
-    </aside>
   );
 }
