+++ packages/react/src/components/layout/index.ts
@@ -1,2 +1,3 @@
 export * from './FocusLayout';
 export * from './GridLayout';
+export { CarouselView } from './CarouselView';
