+++ packages/styles/scss/components/_index.scss
@@ -11,6 +11,7 @@
 // Layout
 @use './layout/focus-layout';
 @use './layout/grid-layout';
+@use './layout/carousel-view';
 
 // Participant
 @use './participant/connection-quality';
